Add ApplicationFrameHost.exe
Starting Edge manually starts this process, so it appears to be part of
the Metro app framework. I've added it to EdgeDbg as I assume that makes
it a "part" of Edge. Even if it's not, it makes sense to want to kill
and restart it to prevent issues in this process from interfering with
launching Edge.