Attempt to suspend processes faster
Rather than wait for processes to start in sequence and suspend them,
wait for any process and suspend in the order in which they are
detected.
Also allow MicrosoftEdgeCP.exe to not have been started yet in order to
reduce the chance of Edge loading content before the debugger has
attached. This means the process id for this process may be 0, which
means it has not been started yet